create BaseTeaserActivity with shared code between Teaser and Replayer
