+++ common/src/main/res/layout/activity_main.xml
@@ -5,8 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/root"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activity.main.MainActivity">
+    android:layout_height="match_parent" >
 
     <com.mapbox.vision.view.VisionView
         android:id="@+id/vision_view"
@@ -129,17 +128,6 @@
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
-
-    <com.mapbox.vision.examples.ScoreProgressView
-        android:id="@+id/driver_score_mode"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:visibility="gone"
-        tools:visibility="visible"
-        android:layout_marginTop="18dp"
-        android:layout_marginEnd="18dp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/dashboard_container"
@@ -414,12 +402,6 @@
                     android:textStyle="bold"/>
 
             </LinearLayout>
-
-            <View
-                android:layout_width="1dp"
-                android:layout_height="match_parent"
-                android:visibility="gone"
-                android:background="@drawable/menu_line_gradient_rotated_bg"/>
 
         </LinearLayout>
