+++ common/src/main/res/values/strings.xml
@@ -11,4 +11,6 @@
     <string name="go">Go</string>
     <string name="tap_to_select_destination">Tap on the map to select a destination</string>
     <string name="can_not_calculate_new_route">Cannot calculate new route</string>
+    <string name="choose_session">Choose Session</string>
+    <string name="start_ar_session">Start AR Session</string>
 </resources>
