+++ replayer/src/main/res/values/strings.xml
@@ -1,5 +1,3 @@
 <resources>
     <string name="app_name">Session Replayer</string>
-    <string name="choose_session">Choose Session</string>
-    <string name="start_ar_session">Start AR Session</string>
 </resources>
